Moduł 5 - Logowanie, generowanie tokena, zabezpieczenia dostępu

+++ WebAPI/Controllers/V1/PostsController.cs
@@ -1,8 +1,10 @@
 ﻿using Application.Dto;
 using Application.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 //using Microsoft.AspNet.OData;
 using Microsoft.AspNetCore.Mvc;
 using Swashbuckle.AspNetCore.Annotations;
+using System.Security.Claims;
 using WebAPI.Filters;
 using WebAPI.Helpers;
 using WebAPI.Wrappers;
@@ -13,6 +15,8 @@ namespace WebAPI.Controllers.V1
     //Zapamiętać, że async zmienić w Repository Services oraz Controllers
     [ApiVersion("1.0")]
     [Route("api/[controller]")]
+    //umozliwa ograniczenie dostepu do zasobów na podstawie ról
+    [Authorize]
     [ApiController]
     public class PostsController : Controller
     {
@@ -24,13 +28,15 @@ namespace WebAPI.Controllers.V1
         }
 
         [SwaggerOperation(Summary = "Retrieves sort fields")]
+        [AllowAnonymous]
         [HttpGet("[action]")]
-        public async Task<IActionResult> GetSortFields()
+        public IActionResult GetSortFields()
         {
             return Ok(SortingHelper.GetSortFields().Select(x => x.Key));
         }
 
         [SwaggerOperation(Summary = "Retrieves all posts")]
+        [AllowAnonymous]
         [HttpGet]
         public async Task<IActionResult> Get([FromQuery] /*Wartość parametru zostanie pobrana z ciągu zapytania*/ PaginationFilter paginationFilter,
                                                     [FromQuery] SortingFilter sortingFilter,
@@ -53,6 +59,7 @@ namespace WebAPI.Controllers.V1
 
         [SwaggerOperation(Summary = "Retrieves paged posts")]
         //[EnableQuery]
+        [AllowAnonymous]
         [HttpGet("[action]")]
         public IQueryable<PostDto> GetAll()
         {
@@ -60,6 +67,7 @@ namespace WebAPI.Controllers.V1
         }
 
         [SwaggerOperation(Summary = "Retrieves a specyfic post by unique id")]
+        [AllowAnonymous]
         [HttpGet("{id}")]
         public async Task<IActionResult> Get(int id)
         {
@@ -75,7 +83,7 @@ namespace WebAPI.Controllers.V1
         [HttpPost]
         public async Task<IActionResult> Create(CreatePostDto newPost)
         {
-            var post = await _postService.AddNewPostAsync(newPost);
+            var post = await _postService.AddNewPostAsync(newPost, User.FindFirstValue(ClaimTypes.NameIdentifier));
             return Created($"api/posts/{post.Id}", new Response<PostDto>(post));
         }
 
@@ -83,6 +91,19 @@ namespace WebAPI.Controllers.V1
         [HttpPut]
         public async Task<IActionResult> Update(UpdatePostDto updatePost)
         {
+            var userOwnPost = await _postService.UserOwnPostAsync(updatePost.Id, User.FindFirstValue(ClaimTypes.NameIdentifier));
+            
+            if (!userOwnPost)
+            {
+                var response = new Response<bool>
+                {
+                    Succeeded = false,
+                    Message = "You do not own this post"
+                };
+
+                return BadRequest(response);
+            }
+                
             await _postService.UpdatePostAsync(updatePost);
             return NoContent();
         }
@@ -91,6 +112,19 @@ namespace WebAPI.Controllers.V1
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            var userOwnPost = await _postService.UserOwnPostAsync(id, User.FindFirstValue(ClaimTypes.NameIdentifier));
+
+            if (!userOwnPost)
+            {
+                var response = new Response<bool>
+                {
+                    Succeeded = false,
+                    Message = "You do not own this post"
+                };
+
+                return BadRequest(response);
+            }
+
             await _postService.DeletePostAsync(id);
             return NoContent();
         }
